==> 07_example_parliament_speech.qmd <==
---
editor: visual
callout-icon: false
format:
  pdf:
    fontsize: "12pt"
    documentclass: report
    classoption: [twocolumn]
    pagestyle: empty
    geometry:
      - top=10mm
      - left=10mm
      - right=10mm
      - bottom=10mm
output-file: "labeled_parliament_speech_2.pdf"
---

Frau Präsidentin! Liebe Kolleginnen und Kollegen! Ich möchte diese Handelsdebatte zu Beginn gerne in die gesamtwirtschaftliche Lage einordnen. Herr Kruse hat ja gesagt, es solle etwas großteiliger werden.

::: callout-tip
## Rezession und Wirtschaftspolitik
Wir haben in dieser Woche die Sachverständigen gehört. Wir sind das einzige Industrieland auf der Welt, das wirtschaftlich schrumpft. Deutschland wird ärmer, die Deutschen werden ärmer – bei einer Rekordinflation in den letzten zwei Jahren. Die Exportnation schwächelt, die Ausfuhren sinken. Jeden Tag werden in Deutschland Investitionsentscheidungen gegen Deutschland getroffen. Wer kann, investiert im Ausland.
:::

::: callout-tip
## Rezession und Wirtschaftspolitik
Das ist keine konjunkturelle Delle, wie es uns hier der Wirtschaftsminister beständig erzählt. Der Standort Deutschland und mit ihm unser Wohlstand sind in Gefahr.
:::

Der Sachverständigenrat sagt: Die mittelfristigen Wachstumsaussichten sind auf historischem Tiefstand.

Liebe Kolleginnen und Kollegen, diejenigen, die das sagen, bezichtigen Sie der Schlechtrederei. Das Gegenteil ist aber das Problem; die Schönredner auf der Regierungsbank sind das Problem.

::: callout-tip
## Rezession und Wirtschaftspolitik
Nehmen Sie endlich wahr, wie die wirtschaftliche Lage im Land ist!
:::

::: callout-note
## Bürgergeld-Reform und Arbeitsmarktintegration
Was zu tun wäre, ist kein Hexenwerk oder keine Raketenwissenschaft. Es ist eigentlich ganz einfach: Wir müssten den Standort stärken, Steuern und Abgaben senken, Bürokratie abbauen, zur Arbeitsaufnahme anreizen, Energiekosten senken und die Handelspolitik stärken. Das Problem ist: Sie machen bei all diesen Themen nicht nur nichts, sondern das Gegenteil von dem, was nötig wäre: mehr Bürokratie – Lieferkettengesetz und Heizungsgesetz –, mehr Steuern und Abgaben – die Sozialversicherungsabgaben sind auf Rekordniveau –, weniger Arbeitsanreize – Bürgergeld hoch und damit für die erwerbsfähigen Bürgergeldempfänger weniger Anreiz zu arbeiten.
:::

::: callout-important
## Erneuerbare Energien und Energiewende

Die Energiekosten werden auch noch erhöht statt gesenkt. Sie waren es doch, die 6 Prozent der deutschen Stromproduktion mitten in der Krise vom Netz genommen haben. Sie haben drei sichere Kernkraftwerke abgeschaltet, die günstig klimaneutralen Strom produziert haben. Sie sind mitverantwortlich dafür, dass die Strompreise in Deutschland so hoch sind, wie sie sind.
:::

::: callout-important
## Erneuerbare Energien und Energiewende

Sie schrauben die Preise hoch. Und hier möchte ich etwas zu dem sagen, was Sie uns gestern mal wieder per Presserklärung statt heute im Bundestag mitgeteilt haben, nämlich wie es weitergehen soll. Sie feiern sich für Ihr Paketchen von gestern. Jetzt muss man sich mal anschauen, was da für die energieintensive Industrie in Deutschland vereinbart worden ist.
:::

::: callout-note
## SPD-internes Handeln und Positionen
– Dass Sie das feiern, ist besonders bemerkenswert angesichts dessen, was Sie als SPD-Fraktion in den letzten sechs Monaten für einen Popanz zum Industriestrompreis gemacht haben.
:::

Für die energieintensive Industrie in Deutschland ändert sich durch das, was Sie vereinbart haben, faktisch nichts. Sie verkaufen das Weglassen von Belastungen, die Sie zum 1. Januar geplant haben, jetzt als Entlastung. Das ist eine Mogelpackung und nichts anderes.

Man muss das mal politisch einordnen: Olaf Scholz und Christian Lindner haben Robert Habeck brutal auflaufen lassen. Sie sind krachend gescheitert,


übrigens auch mit Ihrem Ansatz. Chemieindustrie, Papierindustrie und Stahlindustrie werden weiter aus Deutschland abwandern. Sie als Ampel sind die eigentliche Standortgefahr für Deutschland. Das ist das Problem Ihrer Politik in den letzten Monaten.


::: callout-warning
## Freihandelsabkommen und CETA
Das setzt sich bei der Handelspolitik fort. Wir sehen ein völliges Scheitern Ihrer Handelspolitik. Die Gespräche zu Zollfragen mit den USA sind ohne Ergebnis abgebrochen. Auch die Gespräche mit Australien sind ohne Ergebnis abgebrochen. Von Chefsache keine Spur!
:::

::: callout-warning
## Freihandelsabkommen und CETA
Das Allerschlimmste aber betrifft Mercosur. Man muss die Mercosur-Verhandlungen mal in ihrer Dimension einordnen.
:::

::: callout-warning
## Freihandelsabkommen und CETA
Ein Handelsabkommen der EU mit den Mercosur-Staaten Südamerikas, unter anderem Brasilien und Argentinien, würde mit über 700 Millionen Menschen, die dort leben, den größten Handelsraum der Welt schaffen. Es würde Europa und Südamerika näher zusammenbringen in einer Zeit, in der wir angesichts dessen, was los ist, mehr Partner auf der Welt brauchen. Und was machen Sie? Sie als deutsche Bundesregierung überfrachten das Abkommen mit immer neuen Forderungen, wollen Zusatzerklärungen zu Klima und Umwelt. Sie wollen Sanktionsmechanismen, das heißt nichts anderes als Strafmechanismen. Unsere Partner in Südamerika empfinden das als Misstrauensvotum, als Affront. Und wissen Sie was? Ich kann die Südamerikaner gut verstehen. Alle in Brüssel und Südamerika sagen einhellig: Es ist vor allem diese Bundesregierung, die Mercosur aktiv gefährdet.
:::

Und so muss man es einordnen: Es wäre ein historisches Versagen, wenn dieses wichtige Abkommen zwischen der Europäischen Union und den südamerikanischen Staaten an Ihren Forderungen, an Ihnen scheitert. Das wäre ein historisches Versagen. So muss man es Ihnen ins Stammbuch schreiben.

::: callout-warning
## Ampelkoalition und politische Krise
In der letzten Debatte vor der Sommerpause zu diesem Thema hat die Ampel – ich zitiere Frau Staatssekretärin Brantner – angekündigt: „Wir sind diejenigen, die zeigen, dass uns diese Länder wichtig sind …“ Der Kanzler war da, der Vizekanzler war da, die Außenministerin war da, der Landwirtschaftsminister war da, die Ministerin für wirtschaftliche Zusammenarbeit und Entwicklung war da. Mein Eindruck ist: Genau das ist das Problem.
:::

::: callout-note
## China-Strategie und Partnerschaft
Die Generaldirektorin der Welthandelsorganisation hat im September auf der Botschafterkonferenz in Berlin gesagt – ich zitiere wieder –: „When we talk to China, we get an airport; when we talk to Germany, we get a lecture.“
:::

Sie reisen durch die Welt, halten Reden, belehren unsere Partner, misstrauen ihnen, überfrachten die Handelsverträge und treiben unsere Partner in Chinas Arme. Sie machen eine Handelspolitik in die völlig falsche Richtung.

Unser nationales Interesse als Exportnation ist es, möglichst viele Handelsverträge zu schließen. Wir müssen Autos, Maschinen, unsere Produkte und Innovationen in die Welt verkaufen, damit wir als älter werdende Gesellschaft uns Gesundheit, Pflege und Rente überhaupt noch leisten können. Ihre moralingetränkte Außen- und Handelspolitik vergrätzt unsere Partner und schadet unseren nationalen Interessen. Nutzen Sie die Debatte heute – ich ahne ja schon, dass Sie uns gleich wieder erzählen werden, dass eigentlich alles in Ordnung ist; Schönfärberei! – endlich für eine ehrliche Bestandsaufnahme,

::: callout-warning
## Freihandelsabkommen und CETA
und tun Sie, was nötig ist, um Mercosur zum Erfolg zu bringen!
:::
