==> 08_example_parliament_speech.qmd <==
---
editor: visual
callout-icon: false
format:
  pdf:
    fontsize: "12pt"
    documentclass: report
    classoption: [twocolumn]
    pagestyle: empty
    geometry:
      - top=10mm
      - left=10mm
      - right=10mm
      - bottom=10mm
output-file: "labeled_parliament_speech.pdf"
---

::: callout-tip
## AfD und parlamentarische Prozesse

Sehr geehrte Frau Präsidentin! Verehrte Kolleginnen! Verehrte Kollegen! Meine Damen und Herren! Der Antrag der AfD heißt „Elf-Punkte-Plan“, ist aber am Ende sehr eindimensional. Ihr Antrag ist getrieben von Angst, und Sie verfolgen auch das Ziel, Ängste zu schüren. Angst ist, musikalisch gesprochen, das Leitmotiv, der Basso continuo Ihrer Politik. Entsprechend ist auch Ihr Antrag eindimensional.
:::

Die einzige Idee, die Sie im Endeffekt – wenn man alles summiert und zusammenzählt – zur Migration haben, ist, eine Festung Europa zu bauen, und Ihr Instrument dazu sind im Grunde Grenzkontrollen – womit Sie aber eigentlich Grenzschließungen meinen – und pauschale und unterschiedslose Abschiebungen.

Das Ziel, das Sie verfolgen, ist, einfach ausgedrückt: einfach alle abschieben, und zwar so schnell wie möglich.

Angst ist aber kein guter Ratgeber für eine intelligente Politik.

::: callout-note
## Migrations- und Asylpolitik

Natürlich sind Grenzkontrollen und auch Rückführungen auch eine Dimension von Asyl- und Ausländerpolitik und des Migrationsrechts.
:::

Aber man kann Deutschland, man kann Europa nicht einfach hermetisch abschotten. Wir haben eine Landgrenze nach Osten hin zu Asien, wir haben ein Binnenmeer nach Süden hin zu Afrika. Und auch völkerrechtlich, verfassungsrechtlich und humanitär bleiben am Ende Asylrecht und Flüchtlingsschutz übrig.

::: callout-note
## Migrations- und Asylpolitik

In den meisten Zeiten war das auch immer beherrschbar. Zu Kriegszeiten – Jugoslawien-Krieg, Syrien, jetzt in der Ukraine – schnellen die Zahlen in die Höhe. Aber das ist niemals planbar. Migration ist eben auch etwas Mehrdimensionales; demgemäß ist auch die Migrationspolitik der Koalition mehrdimensional, erfasst die Komplexität in ihrer Gänze.
:::

::: callout-note
## Migrations- und Asylpolitik

Deswegen setzt unsere Neuordnung in der Migrationspolitik nicht nur auf ein Instrument, sondern auf mehrere Instrumente, nicht nur auf Kontrolle und Rückführungsoffensive, sondern die Frage ist auch, wie wir eigentlich mit den Menschen umgehen, die als Asylberechtigte oder als Menschen, die Flüchtlingsschutz erhalten, oder aus anderen Gründen für lange Zeit hierbleiben werden. Sie können natürlich nicht auf Dauer in staatlichen Unterkünften stecken bleiben
:::

::: callout-warning
## Wohnungsbau und Mietpreise

oder im Sozialsystem. Wir sehen und erleben doch alle, wenn wir rausgehen, wie überfüllt die Unterkünfte der Kommunen sind, dass die aus allen Nähten platzen,
:::

zum Teil immer noch belegt sind mit Menschen, die seit Jahren hier sind

::: callout-warning
## Wohnungsbau und Mietpreise

und die wir nicht in den Arbeitsmarkt, in den Wohnungsmarkt hineinbringen.
:::

::: callout-note
## Migrations- und Asylpolitik

Deswegen ist es übrigens auch gut, dass wir über das Chancen-Aufenthaltsrecht, das wir im letzten Jahr noch beschlossen haben,
:::

versuchen, diese Menschen nach Möglichkeit in Arbeit und dann auch möglichst in Wohnungen zu bringen.

::: callout-caution
## Bundeswehr Finanzierung und Ausstattung

Genau deshalb ist es eindimensional, zu kurz gedacht, zu einfach, unterkomplex gedacht, alle Ressourcen nur in Abwehr und Abschottung zu stecken. Das hat uns all die Jahre nicht weitergebracht.
:::

Wir brauchen auch einen Teil der Ressourcen für Qualifikation und Integration. Denn die meisten Migranten wollen doch etwas erreichen, wollen etwas für sich erreichen, wollen es zu etwas bringen. Natürlich: Wer was erreichen will, muss sich auch anstrengen im Leben.

Klar, wer Gutes will, ist immer fein raus. Aber man tut den Menschen nichts Gutes, wenn man ihnen Anstrengungen erspart.

Deswegen muss man auch sagen: Wir wollen ein Aufstiegsversprechen geben, auch an zugewanderte Menschen. Aber es ist ein Versprechen an Menschen, die auch ihren Beitrag erbringen wollen. Deswegen müssen wir auch mehr investieren in die Möglichkeiten, in Angebote für Bildung, Ausbildung, Fortbildung und Weiterbildung. Aber das müssen wir auch abverlangen. Das ist keine Zumutung, keine Körperverletzung; es gehört einfach dazu, diese Anstrengungen zu verlangen.

Deswegen müssen wir auch wegkommen von dem Bild, dass Migranten nur Bittsteller sind. Das ist das Bild, das Sie im Kopf haben und von dem Sie wollen, dass es sich verstärkt. Wir wollen, dass Sie von dieser Eindimensionalität wegkommen und die zweite Dimension sehen, die ganze Komplexität.

Wer hier ist und wer hier bleibt, hat Rechte, aber eben auch Pflichten – wie jeder andere, der in diesem Land lebt.

::: callout-note
## Migrations- und Asylpolitik

Dass wir, ja, auch Teile des Asyl- und Ausländerrechts nicht konsequent vollziehen, ist schlecht. Aber es ist eben auch schlecht, dass wir bei Integration und Qualifikation noch nicht gut genug sind. Wir müssen besser werden – wir werden übrigens morgen in diesem Hause auch präsentieren, wie wir uns solche Dinge vorstellen –;
:::

denn das wäre rechtlich, wirtschaftlich und kulturell wahnsinnig wichtig für alle. Wir würden – und damit komme ich zum Schluss, Frau Präsidentin – dann nicht nur unsere humanitären, völker- und verfassungsrechtlichen Pflichten erfüllen, sondern wir würden auch mehr für die volkswirtschaftlichen und betrieblichen Bedürfnisse im Land tun und auch die gesellschaftliche Akzeptanz im Lande erhöhen. Das ist die ganze Dimension, und die liegt uns am Herzen.
